ui air quality index

+++ app/src/main/java/com/capstone/hibykes/data/remote/response/AirPollutionResponse.kt
@@ -38,7 +38,7 @@ data class Components(
 	val no2: Double? = null,
 
 	@field:SerializedName("no")
-	val no: Int? = null,
+	val no: Double? = null,
 
 	@field:SerializedName("o3")
 	val o3: Double? = null,
